fix: dependency error
