+++ src/components/contact-admin-mailto/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { MailtoLink } from '@openedx/paragon';
 import PropTypes from 'prop-types';
 import { getContactEmail } from '../../utils/common';
-import { useEnterpriseCustomer } from "../hooks";
+import { useEnterpriseCustomer } from '../hooks';
 
 const ContactAdminMailto = ({
   children,
