--- src/components/enterprise-subsidy-requests/constants.js
@@ -1,12 +1,0 @@
-export const SUBSIDY_TYPE = {
-  LICENSE: 'license',
-  COUPON: 'coupon',
-};
-export const SUBSIDY_REQUEST_STATE = {
-  REQUESTED: 'requested',
-  PENDING: 'pending',
-  APPROVED: 'approved',
-  DECLINED: 'declined',
-  ERROR: 'error',
-};
-export const LOADING_SCREEN_READER_TEXT = 'loading your subsidy requests from your organization';
