+++ src/components/enterprise-subsidy-requests/data/service.js
@@ -1,6 +1,7 @@
 import { getAuthenticatedHttpClient } from '@edx/frontend-platform/auth';
 import { getConfig } from '@edx/frontend-platform/config';
-import { SUBSIDY_REQUEST_STATE } from '../constants';
+
+import { SUBSIDY_REQUEST_STATE } from '../../../constants';
 
 export function fetchSubsidyRequestConfiguration(enterpriseUUID) {
   const url = `${getConfig().ENTERPRISE_ACCESS_BASE_URL}/api/v1/customer-configurations/${enterpriseUUID}/`;
