+++ src/components/hooks/index.js
@@ -1,4 +1,5 @@
 export { default as useEnterpriseCustomer } from './useEnterpriseCustomer';
+export { default as useEntepriseFeatures } from './useEnterpriseFeatures';
 export { default as useSubscriptionLicenses } from './useSubscriptionLicenses';
 export { default as useActiveRedeemablePolicies } from './useActiveRedeemablePolicies';
 export { default as useCouponCodes } from './useCouponCodes';
