+++ src/components/hooks/useEnterpriseFeatures.js
@@ -1,0 +1,8 @@
+import { useEnterpriseLearner } from '../app/data';
+
+export default function useEnterpriseFeatures() {
+  const { data: { enterpriseFeatures } } = useEnterpriseLearner();
+  return {
+    ...enterpriseFeatures,
+  };
+}
