+++ src/components/my-career/SkillsRecommendationCourses.jsx
@@ -1,5 +1,5 @@
 import React, {
-  useContext, useMemo, useState, useEffect,
+  useMemo, useState, useEffect,
 } from 'react';
 import { useNavigate } from 'react-router-dom';
 import PropTypes from 'prop-types';
