+++ src/components/pathway/SearchPathwayCard.jsx
@@ -1,8 +1,7 @@
-import React, { useContext, useMemo } from 'react';
+import React, { useMemo } from 'react';
 import PropTypes from 'prop-types';
 import cardFallbackImg from '@edx/brand/paragon/images/card-imagecap-fallback.png';
 import { useNavigate } from 'react-router-dom';
-import { AppContext } from '@edx/frontend-platform/react';
 import { getConfig } from '@edx/frontend-platform/config';
 import { camelCaseObject } from '@edx/frontend-platform/utils';
 import {
@@ -16,8 +15,7 @@ import {
   PATHWAY_SKILL_QUIZ_EVENT_NAME,
 } from './constants';
 import PathwayModal from './PathwayModal';
-import { useEnterpriseLearner } from '../app/data';
-import { useEnterpriseCustomer } from "../hooks";
+import { useEnterpriseCustomer } from '../hooks';
 
 // This function is for filtering list of skillNames in a way that returning list
 // can be displayed in the form of 2 rows at max.
