+++ src/components/skills-quiz/SkillsQuiz.jsx
@@ -1,8 +1,6 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { Helmet } from 'react-helmet';
 import { SearchData } from '@edx/frontend-enterprise-catalog-search';
-
-import { AppContext } from '@edx/frontend-platform/react';
 
 import { Container, Row } from '@openedx/paragon';
 
@@ -13,7 +11,7 @@ import SkillsQuizStepper from './SkillsQuizStepper';
 import { SkillsContextProvider } from './SkillsContextProvider';
 import SkillsQuizV2 from '../skills-quiz-v2/SkillsQuiz';
 import { isExperimentVariant } from '../../utils/optimizely';
-import { useEnterpriseCustomer } from "../hooks";
+import { useEnterpriseCustomer } from '../hooks';
 
 const SkillsQuiz = () => {
   const enterpriseCustomer = useEnterpriseCustomer();
