+++ src/constants/index.js
@@ -1,1 +1,2 @@
 export * from './course';
+export * from './subsidies';
